+++ src/sentry/receivers/rules.py
@@ -1,6 +1,7 @@
+from sentry import features
+from sentry.models.project import Project
 from sentry.models.rule import Rule
 from sentry.notifications.types import FallthroughChoiceType
-from sentry.rules.conditions.high_priority_issue import has_high_priority_issue_alerts
 from sentry.signals import project_created
 
 DEFAULT_RULE_LABEL = "Send a notification for new issues"
@@ -34,6 +35,22 @@ DEFAULT_RULE_DATA_NEW = {
     ],
     "actions": DEFAULT_RULE_ACTIONS_NEW,
 }
+
+
+PLATFORMS_WITH_NEW_DEFAULT = ["python", "javascript"]
+
+
+def has_high_priority_issue_alerts(project: Project) -> bool:
+    # High priority alerts are enabled if the project has the feature flag
+    # or for python/javascript projects in organization that have the feature flag
+    return features.has("projects:high-priority-alerts", project) or (
+        features.has("organizations:default-high-priority-alerts", project.organization)
+        and project.platform is not None
+        and any(
+            project.platform.startswith(base_platform)
+            for base_platform in PLATFORMS_WITH_NEW_DEFAULT
+        )
+    )
 
 
 def create_default_rules(project, default_rules=True, RuleModel=Rule, **kwargs):
