+++ src/sentry/rules/conditions/high_priority_issue.py
@@ -1,33 +1,16 @@
 from datetime import datetime
 from typing import Optional, Sequence
 
-from sentry import features
 from sentry.eventstore.models import GroupEvent
 from sentry.models.activity import Activity
 from sentry.models.group import Group
-from sentry.models.project import Project
+from sentry.receivers.rules import has_high_priority_issue_alerts
 from sentry.rules import EventState
 from sentry.rules.conditions.base import EventCondition
 from sentry.types.activity import ActivityType
 from sentry.types.condition_activity import ConditionActivity, ConditionActivityType
 
 HIGH_SEVERITY_THRESHOLD = 0.1
-
-
-PLATFORMS_WITH_NEW_DEFAULT = ["python", "javascript"]
-
-
-def has_high_priority_issue_alerts(project: Project) -> bool:
-    # High priority alerts are enabled if the project has the feature flag
-    # or for python/javascript projects in organization that have the feature flag
-    return features.has("projects:high-priority-alerts", project) or (
-        features.has("organizations:default-high-priority-alerts", project.organization)
-        and project.platform is not None
-        and any(
-            project.platform.startswith(base_platform)
-            for base_platform in PLATFORMS_WITH_NEW_DEFAULT
-        )
-    )
 
 
 class HighPriorityIssueCondition(EventCondition):
